% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostBiCopPDF.R
\name{boostBiCopPDF}
\alias{boostBiCopPDF}
\title{Density of a Conditional Bivariate Copula}
\usage{
boostBiCopPDF(object, U, X, ...)
}
\arguments{
\item{object}{an \code{boostBiCop} object.}

\item{U}{a matrix with two columns with values in [0,1].}

\item{X}{matrix containing the covariates and having the same number of rows as \code{U}.}

\item{...}{unused.}
}
\value{
A numeric vector containing the density values of the considered \code{boostBiCop} object, evaluated at \code{U} and \code{X}.
}
\description{
Probability density function (PDF) of an \code{boostBiCop} object.
}
\examples{
# load simulated data
data(data_bicop)

# fit object
object <- boostBiCopEst(formula = ~.,
                        U = data_bicop[, 1:2],
                        X = data_bicop[, -c(1:2)],
                        family = 301,
                        control = boost_control(deselection = "attributable"))
boostBiCopPDF(object,
              U = data_bicop[, 1:2],
              X = data_bicop[, -c(1:2)])

}
