% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineCopSelect.R
\name{boostRVineCopSelect}
\alias{boostRVineCopSelect}
\title{Sequential Selection and Gradient-Boosted Estimation of GLMs for a Conditional R-Vine Copula Model}
\usage{
boostRVineCopSelect(
  U,
  X,
  boostRVM,
  familyset = NA,
  selectioncrit = "aic",
  indeptest = NA,
  trunclevel = NA,
  control = boost_control(),
  na.action = na.omit,
  light = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{U}{a matrix with d columns with values in [0,1].}

\item{X}{matrix containing the covariates and having the same number of rows as \code{U}.}

\item{boostRVM}{An \code{\link{boostRVineMatrix}} object.
Note that the \code{family} matrix in the \code{boostRVineMatrix} object will be ignored to select from the \code{familyset}.}

\item{familyset}{vector of bivariate copula families as described in \code{\link{boostBiCopEst}} to select from. If \code{NA} (default),
selection among all implemented copula families is performed. Note, that a defined \code{family} matrix in \code{boostRVM} will be ignored
and therefore does not need to be specified.}

\item{selectioncrit}{character; either "\code{loglik}" or "\code{aic}" (default) for the copula family selection.}

\item{indeptest}{test for independence for each edge in the regular vine. Default: \code{NA}, i.e no independence check.
Otherwise, numeric value specifies significance level. The independence copula is chosen if the null hypothesis of independence cannot be rejected.}

\item{trunclevel}{integer; level of truncation.}

\item{control}{control parameters for the gradient-boosting estimation.
The \code{\link{boost_control}} function provides the default setting.}

\item{na.action}{a function which indicates what should happen if the data contains \code{NA}s. Default: \code{na.omit}.}

\item{light}{logical; if \code{FALSE} (default) the model frame is kept in the \code{boostBiCop} objects.
Otherwise, if \code{TRUE}, the model frame is omitted.}

\item{cores}{integer; the number of cores used for computations. Default: \code{cores = 1}.}

\item{...}{unused.}
}
\value{
An \code{boostRVineCopula} object.
}
\description{
This function sequentially estimates GLMs for conditional bivariate copulas of a d-dimensional conditional R-vine copula model
as specified by the corresponding \code{\link{boostRVineMatrix}} object via gradient-boosting.
Additionally, it selects the best fitting copula family for each conditional bivariate copula.
}
\note{
The code of this function is mainly based on the R-package \code{\link{VineCopula}} by Nagler et al. but slightly adapted to the needs of this package.
}
\examples{
# load simulated data
data(data_vinecop)

# create boostRVineMatrix
Matrix <- c(5, 0, 0, 0, 0,
            2, 2, 0, 0, 0,
            3, 3, 3, 0, 0,
            1, 4, 4, 4, 0,
            4, 1, 1, 1, 1)
Matrix <- matrix(Matrix, 5, 5, byrow = TRUE)
Family <- matrix(0, nrow = 5, ncol = 5)
Formula <- matrix("~ .", 5, 5, byrow = TRUE)
boostRVM <- boostRVineMatrix(Matrix = Matrix,
                             family = Family,
                             formula = Formula)

# fit object
object <- boostRVineCopSelect(U = data_vinecop[, 1:5],
                              X = data_vinecop[, -c(1:5)],
                              boostRVM = boostRVM,
                              familyset = c(1, 301, 304, 401, 404),
                              selectioncrit = "aic",
                              control = boost_control(deselection = "attributable"),
                              cores = 1)
summary(object)


}
