% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineMatrix.R
\name{boostRVineMatrix}
\alias{boostRVineMatrix}
\title{Conditional R-Vine Copula Matrix Notation}
\usage{
boostRVineMatrix(
  Matrix,
  family = array(NA, dim = dim(Matrix)),
  formula = array(NA, dim = dim(Matrix))
)
}
\arguments{
\item{Matrix}{Lower triangular d x d integer matrix that defines the R-vine tree structure.}

\item{family}{Lower triangular d x d integer matrix that assigns the pair-copula families to each (conditional) pair defined by Matrix
(default: \code{family = array(NA, dim = dim(Matrix))}). The available bivariate copula families are given in \code{\link{boostBiCopEst}}.}

\item{formula}{Lower triangular d x d character matrix that assigns the formula to the corresponding copula family specified in \code{family}
(default: \code{formula = array(NA, dim = dim(Matrix))}).}
}
\value{
An \code{boostRVineMatrix} object.
}
\description{
This function creates an \code{boostRVineMatrix} object for a d-dimensional conditional R-vine copula
containing the R-vine structure, the specified pair-copula families and the corresponding formulas.
}
