% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineSeqEst.R
\name{boostRVineSeqEst}
\alias{boostRVineSeqEst}
\title{Sequential Gradient-Boosted Estimation of GLMs for a Conditional R-Vine Copula Model}
\usage{
boostRVineSeqEst(
  U,
  X,
  boostRVM,
  control = boost_control(),
  na.action = na.omit,
  light = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{U}{a matrix with d columns with values in [0,1].}

\item{X}{matrix containing the covariates and having the same number of rows as \code{U}.}

\item{boostRVM}{An \code{\link{boostRVineMatrix}} object including the R-vine structure in \code{Matrix},
the specified pair-copula families in \code{family} and the corresponding formulas in \code{formula}.}

\item{control}{control parameters for the gradient-boosting estimation.
The \code{\link{boost_control}} function provides the default setting.}

\item{na.action}{a function which indicates what should happen if the data contains \code{NA}s. Default: \code{na.omit}.}

\item{light}{logical; if \code{FALSE} (default) the model frame is kept in the \code{boostBiCop} objects.
Otherwise, if \code{TRUE}, the model frame is omitted.}

\item{cores}{integer; the number of cores used for computations. Default: \code{cores = 1}.}

\item{...}{unused.}
}
\value{
An \code{boostRVineCopula} object.
}
\description{
This function sequentially estimates the GLMs of conditional bivariate copulas of a d-dimensional conditional R-vine copula model
as specified by the corresponding \code{\link{boostRVineMatrix}} object via gradient-boosting.
}
\note{
The code of this function is mainly based on the R-package \code{\link{VineCopula}} by Nagler et al. but slightly adapted to the needs of this package.
}
\examples{
# load simulated data
data(data_vinecop)

# create boostRVineMatrix
Matrix <- c(5, 0, 0, 0, 0,
            2, 2, 0, 0, 0,
            3, 3, 3, 0, 0,
            1, 4, 4, 4, 0,
            4, 1, 1, 1, 1)
Family <- c(0, 0, 0, 0, 0,
            404, 0, 0, 0, 0,
            1, 404, 0, 0, 0,
            401, 301, 304, 0, 0,
            1, 401, 304, 301, 0)
Matrix <- matrix(Matrix, 5, 5, byrow = TRUE)
Family <- matrix(Family, nrow = 5, ncol = 5, byrow = TRUE)
Formula <- matrix("~ .", 5, 5, byrow = TRUE)
boostRVM <- boostRVineMatrix(Matrix = Matrix,
                             family = Family,
                             formula = Formula)

# fit object
object <- boostRVineSeqEst(U = data_vinecop[, 1:5],
                           X = data_vinecop[, -c(1:5)],
                           boostRVM = boostRVM,
                           control = boost_control(deselection = "attributable"),
                           cores = 1)
summary(object)

}
