% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineSim.R
\name{boostRVineSim}
\alias{boostRVineSim}
\title{Simulation from a Conditional R-Vine Copula Model}
\usage{
boostRVineSim(boostRVC, U, X)
}
\arguments{
\item{boostRVC}{an \code{boostRVineCopula} object to sample from.}

\item{U}{a matrix with d columns containing (randomly) uniformly distributed values in [0,1].}

\item{X}{matrix containing the covariates and having the same number of rows as \code{U}.}
}
\value{
A matrix with d columns containing the simulated data from the given d-dimensional conditional R-vine copula.
}
\description{
This function simulates from an \code{boostRVineCopula} object.
}
\note{
The code of this function is mainly based on the R-package \code{\link{VineCopula}} by Nagler et al. but slightly adapted to the needs of this package.
}
\examples{
# load simulated data
data(data_vinecop)

# create boostRVineMatrix
Matrix <- c(5, 0, 0, 0, 0,
            2, 2, 0, 0, 0,
            3, 3, 3, 0, 0,
            1, 4, 4, 4, 0,
            4, 1, 1, 1, 1)
Family <- c(0, 0, 0, 0, 0,
            404, 0, 0, 0, 0,
            1, 404, 0, 0, 0,
            401, 301, 304, 0, 0,
            1, 401, 304, 301, 0)
Matrix <- matrix(Matrix, 5, 5, byrow = TRUE)
Family <- matrix(Family, nrow = 5, ncol = 5, byrow = TRUE)
Formula <- matrix("~ .", 5, 5, byrow = TRUE)
boostRVM <- boostRVineMatrix(Matrix = Matrix,
                             family = Family,
                             formula = Formula)

# fit object
object <- boostRVineSeqEst(U = data_vinecop[, 1:5],
                           X = data_vinecop[, -c(1:5)],
                           boostRVM = boostRVM,
                           control = boost_control(deselection = "attributable"),
                           cores = 1)
boostRVineSim(object,
              U = data_vinecop[, 1:5],
              X = data_vinecop[, -c(1:5)])

}
