% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineSimStudy.R
\name{boostRVineSimStudy}
\alias{boostRVineSimStudy}
\title{Simulation from an Artificial Conditional R-Vine Copula Model}
\usage{
boostRVineSimStudy(boostRVC, U, response)
}
\arguments{
\item{boostRVC}{an \code{boostRVineCopula} object to sample from.}

\item{U}{a matrix with d columns containing (randomly) uniformly distributed values in [0,1].}

\item{response}{vector containing the linear predictor assumed among all bivariate conditional copulas.}
}
\value{
A matrix with d columns containing the simulated data from the given d-dimensional conditional R-vine copula.
}
\description{
This function simulates from a given d-dimensional artificial conditional R-vine copula model.
}
\note{
The code of this function is mainly based on the R-package \code{\link{VineCopula}} by Nagler et al. but slightly adapted to the needs of this package.
}
