% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostRVineTrueLL.R
\name{boostRVineTrueLL}
\alias{boostRVineTrueLL}
\title{Calculate true log-likelihood of an Artificial Conditional R-Vine Copula Model}
\usage{
boostRVineTrueLL(boostRVM, U, response, cores = 1)
}
\arguments{
\item{boostRVM}{An \code{\link{boostRVineMatrix}} object including the R-vine structure in \code{Matrix},
the specified pair-copula families in \code{family} and the corresponding formulas in \code{formula}.}

\item{U}{a matrix with d columns with values in [0,1].}

\item{response}{vector containing the linear predictor assumed among all bivariate conditional copulas.}

\item{cores}{integer; the number of cores used for computations. Default: \code{cores = 1}.}
}
\value{
An \code{boostRVineCopula} object.
}
\description{
Calculate true log-likelihood of an Artificial Conditional R-Vine Copula Model
}
\note{
The code of this function is mainly based on the R-package \code{\link{VineCopula}} by Nagler et al. but slightly adapted to the needs of this package.
}
