% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_data.R
\docType{data}
\name{data_bicop}
\alias{data_bicop}
\title{Simulated data from conditional bivariate copula}
\format{
An object of class \code{data.frame} with 1000 rows and 102 columns.
}
\usage{
data(data_bicop)
}
\description{
The data object 'data_bicop' contains copula data in the first two columns and covariates in the following 100 columns.
The copula data is simulated from copula family 301 with conditional copula model tau = tanh(0.1 - 0.2X1 + 0.3X2 + 0.2X3 + 0.5X4 - 0.4X5).
}
\examples{
data(data_bicop)
}
\keyword{datasets}
