% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinecop_data.R
\docType{data}
\name{data_vinecop}
\alias{data_vinecop}
\title{Simulated data from 5-dimensional conditional regular vine copula}
\format{
An object of class \code{data.frame} with 1000 rows and 105 columns.
}
\usage{
data(data_vinecop)
}
\description{
The data object 'data_vinecop' contains copula data in the first five columns and covariates in the following 100 columns.
The copula data is simulated from a regular vine copula with conditional copula model tau = tanh(0.1 - 0.2X1 + 0.3X2 + 0.2X3 + 0.5X4 - 0.4X5).
}
\examples{
data(data_vinecop)
}
\keyword{datasets}
