% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boostBiCop.R
\name{plot.boostBiCop}
\alias{plot.boostBiCop}
\title{Plot Coefficient Paths of boostBiCop object.}
\usage{
\method{plot}{boostBiCop}(x, col, loglik = FALSE, ...)
}
\arguments{
\item{x}{an x of class \code{boostBiCop}.}

\item{col}{character; colour(s) for the coefficient paths.}

\item{loglik}{logical; either \code{FALSE} (default), then the coefficient paths are plotted.
Otherwise, if \code{TRUE}, then the log-likelihood contribution paths are plotted.}

\item{...}{unused.}
}
\description{
Plot Coefficient Paths of boostBiCop object.
}
\examples{
# see boostBiCopEst

}
