% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.boostBiCop.R
\name{predict.boostBiCop}
\alias{predict.boostBiCop}
\title{Predict from Conditional Bivariate Copula}
\usage{
\method{predict}{boostBiCop}(object, newdata, type = "parameter", ...)
}
\arguments{
\item{object}{an object of class \code{boostBiCop}.}

\item{newdata}{data frame of covariates for which to predict.}

\item{type}{character; either "\code{parameter}" (default) for the copula parameter,
"\code{tau}" for Kendall's Tau or "\code{response}" for the linear predictor.}

\item{...}{unused.}
}
\value{
a data frame.
}
\description{
Predict from Conditional Bivariate Copula
}
\examples{
# load simulated data
data(data_bicop)

# fit object
object <- boostBiCopEst(formula = ~.,
                        U = data_bicop[, 1:2],
                        X = data_bicop[, -c(1:2)],
                        family = 301,
                        control = boost_control(deselection = "attributable"))
predict(object, type = "parameter")
predict(object, newdata = data_bicop, type = "tau")

}
